+++ src/test/kotlin/calculator/StringCalculatorTest.kt
@@ -1,4 +1,4 @@
-package study
+package calculator
 
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatIllegalArgumentException
